workflow barrel track fixes for o2-eve (#7144)
* workflow barrel track fixes

* workflow barrel track fixes

* clang

* clang

+++ EventVisualisation/DataConverter/src/VisualisationCluster.cxx
@@ -24,12 +24,12 @@ namespace o2
 namespace event_visualisation
 {
 
-VisualisationCluster::VisualisationCluster(double XYZ[])
+VisualisationCluster::VisualisationCluster(float XYZ[])
 {
   setCoordinates(XYZ);
 }
 
-void VisualisationCluster::setCoordinates(double xyz[3])
+void VisualisationCluster::setCoordinates(float xyz[3])
 {
   for (int i = 0; i < 3; i++) {
     mCoordinates[i] = xyz[i];
